清理代码

+++ Prism/src/Prism/Renderer/Renderer2D.cpp
@@ -173,7 +173,7 @@ namespace Prism
 
     void Renderer2D::EndScene()
     {
-        uint32_t dataSize = (uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase;
+        uint32_t dataSize = (uint32_t)((uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase);
         if (dataSize)
         {
             s_Data.QuadVertexBuffer->SetData(s_Data.QuadVertexBufferBase, dataSize);
